+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+### 1.2.1
+
+- Added `logoHeightPx` and removed `logoMaxHeightPx`.
+
+### 1.2.0
+
+- Added a `custom.scss` file to the default theme to allow for simple theme customizations to easily get ohter CSS updates.
+- Added a `logoMaxHeightPx` option to set the logo's `max-height` property.
+- Added `scrollPaddingTopPx` option to set the `scroll-padding-top` property for when there's a Nav bar situation.
+
+### 1.1.5
+
+- CSS will be concatenated with SpectaQL's css coming before other CSS. https://github.com/anvilco/spectaql/pull/353
+
+### 1.1.4
+
+- Re-published from borked publish of `1.1.3`
+
+### 1.1.3
+
+- Updated dependencies.
+- Fixed bug where List return type for Query or Mutation was not working properly. https://github.com/anvilco/spectaql/pull/342
+- Fixed navigation bug. https://github.com/anvilco/spectaql/pull/343
+
 ### 1.1.2
 
 - Fixed mismatched header tags